[FIX] payment_stripe: merge stripe checkout and manage forms
Due to the complication importing different files for version saas-16.4
in internal we decided to be consistent with the other versions and
keep only one file payment_form.js instead of checkout_form.js and
manage_form.js in stripe.

opw-3590564
