+++ addons/payment_stripe/__manifest__.py
@@ -18,10 +18,8 @@
     'uninstall_hook': 'uninstall_hook',
     'assets': {
         'web.assets_frontend': [
-            'payment_stripe/static/src/js/checkout_form.js',
             'payment_stripe/static/src/js/express_checkout_form.js',
-            'payment_stripe/static/src/js/manage_form.js',
-            'payment_stripe/static/src/js/stripe_mixin.js',
+            'payment_stripe/static/src/js/payment_form.js',
             'payment_stripe/static/src/js/stripe_options.js',
         ],
     },
